+++ src/Exception/BadMethodCallException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JsonMachine\Exception;
+
+class BadMethodCallException extends JsonMachineException
+{
+}